Consolidated the disk_reader, udp_capture, and udp_sniffer modules under packet_capture.  Cleaned up the associated Python module.
