+++ src/packet_capture.cpp
@@ -216,6 +216,77 @@ BFpacketcapture_status BFpacketcapture_impl::recv() {
 	return ret;
 }
 
+BFstatus bfDiskReaderCreate(BFpacketcapture* obj,
+                            const char*      format,
+                            int              fd,
+                            BFring           ring,
+                            BFsize           nsrc,
+                            BFsize           src0,
+                            BFsize           buffer_ntime,
+                            BFsize           slot_ntime,
+                            BFpacketcapture_callback sequence_callback,
+                            int              core) {
+    return BFpacketcapture_create(obj,
+                                  format,
+                                  fd,
+                                  ring,
+                                  nsrc,
+                                  src0,
+                                  buffer_ntime,
+                                  slot_ntime,
+                                  sequence_callback,
+                                  core,
+                                  BF_IO_DISK);
+}
+
+BFstatus bfUdpCaptureCreate(BFpacketcapture* obj,
+                            const char*      format,
+                            int              fd,
+                            BFring           ring,
+                            BFsize           nsrc,
+                            BFsize           src0,
+                            BFsize           max_payload_size,
+                            BFsize           buffer_ntime,
+                            BFsize           slot_ntime,
+                            BFpacketcapture_callback sequence_callback,
+                            int              core) {
+    return BFpacketcapture_create(obj,
+                                  format,
+                                  fd,
+                                  ring,
+                                  nsrc,
+                                  src0,
+                                  buffer_ntime,
+                                  slot_ntime,
+                                  sequence_callback,
+                                  core,
+                                  BF_IO_UDP);
+}
+
+BFstatus bfUdpSnifferCreate(BFpacketcapture* obj,
+                            const char*      format,
+                            int              fd,
+                            BFring           ring,
+                            BFsize           nsrc,
+                            BFsize           src0,
+                            BFsize           max_payload_size,
+                            BFsize           buffer_ntime,
+                            BFsize           slot_ntime,
+                            BFpacketcapture_callback sequence_callback,
+                            int              core) {
+    return BFpacketcapture_create(obj,
+                                  format,
+                                  fd,
+                                  ring,
+                                  nsrc,
+                                  src0,
+                                  buffer_ntime,
+                                  slot_ntime,
+                                  sequence_callback,
+                                  core,
+                                  BF_IO_SNIFFER);
+}
+
 BFstatus bfPacketCaptureDestroy(BFpacketcapture obj) {
 	BF_ASSERT(obj, BF_STATUS_INVALID_HANDLE);
 	delete obj;
